Updated unit test targets and mock var names
